requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pyhanko[opentype]>=0.21.0
pyhanko-certvalidator>=0.26.0
reportlab>=4.0.0
pypdf>=4.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
cryptography>=41.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
requests>=2.31.0
pytz>=2023.3
